Moved JobStatus enum to types.ts
